--- .npmignore
@@ -1,4 +1,0 @@
-sample
-spec
-*.tgz
-
