+++ package.json
@@ -2,7 +2,11 @@
   "name": "xlf-translate",
   "version": "0.0.0-semantically-released",
   "description": "Populate XLIFF (.XLF) files with existing translations",
-  "main": "translate.js",
+  "main": "src/translate.js",
+  "files": [
+    "bin",
+    "src"
+  ],
   "scripts": {
     "lint": "eslint *.js bin/*",
     "test": "jasmine"
